requirements.txt
streamlit
google-generativeai
python-dotenv
pytesseract
pdf2image
pillow
